pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "form-filler"
description = "CrewAI-powered multi-agent system for processing documents and automatically filling DOCX forms"
readme = "README.md"
authors = [
    {name = "Dat Nguyen", email = "datnguyen.it09@gmail.com"},
]
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
dynamic = ["version"]
keywords = [
    "vietnamese",
    "document-processing",
    "crewai",
    "ollama",
    "ocr",
    "translation",
    "ai-agents",
    "multi-agent-system"
]
dependencies = [
    "click<9.0.0,>=8.0.0",
    "aiohttp<4.0.0,>=3.8.0",

    # CrewAI and AI dependencies
    "crewai<1.0.0,>=0.36.0",
    "crewai-tools<1.0.0,>=0.4.0",
    "langchain<1.0.0,>=0.1.0",
    "langchain-community<1.0.0,>=0.0.20",
    "langchain-ollama<1.0.0,>=0.1.0",

    # Document processing
    "PyMuPDF<2.0.0,>=1.23.0",
    "python-docx<1.0.0,>=0.8.11",
    "mammoth<2.0.0,>=1.5.0",
    "Pillow<10.0.0,>=9.0.0",
    "pytesseract<1.0.0,>=0.3.10",

    # Data handling
    "pydantic<3.0.0,>=2.0.0",
    "python-multipart<1.0.0,>=0.0.5",

    # Utilities
    "colorama<1.0.0,>=0.4.6",
    "tqdm<5.0.0,>=4.64.0",
    "rich<14.0.0,>=13.0.0",
    "loguru<1.0.0,>=0.7.0",

    # CrewAI extensions
    "chromadb<1.0.0,>=0.4.15",
    "sentence-transformers<3.0.0,>=2.2.2",
]

[project.urls]
Homepage = "https://github.com/il-dat/form-filler"
Documentation = "https://form-filler.readthedocs.io"
Repository = "https://github.com/il-dat/form-filler.git"
Issues = "https://github.com/il-dat/form-filler/issues"
Changelog = "https://github.com/il-dat/form-filler/blob/main/CHANGELOG.md"

[project.scripts]
form-filler = "form_filler.cli:main"
form-filler-batch = "form_filler.batch:main"

[project.optional-dependencies]
web = [
    "streamlit<2.0.0,>=1.25.0",
    "plotly<6.0.0,>=5.17.0",
    "pandas<2.3.0,>=2.0.0",
]
dev = [
    "pytest<8.0.0,>=7.0.0",
    "pytest-asyncio<1.0.0,>=0.21.0",
    "pytest-cov<5.0.0,>=4.0.0",
    "pytest-beartype<1.0.0,>=0.0.2",
    "black<24.0.0,>=23.0.0",
    "isort<6.0.0,>=5.12.0",
    "mypy<2.0.0,>=1.5.0",
    "pre-commit<5.0.0,>=4.2.0",
    "beartype<1.0.0,>=0.18.5",
    "ruff>=0.3.0,<1.0.0",
    "poethepoet<1.0.0,>=0.25.0",
]
monitoring = [
    "psutil<6.0.0,>=5.9.0",
    "prometheus-client<1.0.0,>=0.17.0",
]
# docs section removed completely

[tool.pdm.build]
includes = ["src/form_filler"]

[tool.pdm.version]
source = "scm"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/form_filler"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
    "setup.py",
    "demo.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 99
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | temp
  | examples
)/
'''

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "crewai.*",
    "crewai_tools.*",
    "fitz",
    "pytesseract",
    "mammoth",
    "streamlit.*",
    "plotly.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
# E: pycodestyle errors
# F: pyflakes
# I: isort
# N: pep8-naming
# D: pydocstyle
# UP: pyupgrade
# B: flake8-bugbear
# S: flake8-bandit (security)
# C4: flake8-comprehensions
# SIM: flake8-simplify
# TCH: flake8-type-checking
select = ["E", "F", "I", "N", "D", "UP", "B", "S", "C4", "SIM", "TCH"]
ignore = [
    "E501",  # Line too long (handled by black)
    "D100",  # Missing docstring in public module
    "D104",  # Missing docstring in public package
    "D203",  # One blank line required before class docstring
    "D212",  # Multi-line docstring summary should start at the first line
    "D213",  # Multi-line docstring summary should start at the second line
]
typing-modules = ["beartype.typing"]

[tool.ruff.lint.isort]
known-first-party = ["form_filler"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "D103"]  # Allow assert and missing docstrings in tests
"**/__init__.py" = ["F401"]   # Allow unused imports in __init__.py

[tool.bandit]
targets = ["src/"]
exclude = ["tests/", "scripts/", "src/form_filler/utils/"]
skips = ["B101", "B404", "B603"]

[tool.poe.tasks]
test = "pytest tests/"
test-cov = "pytest --cov=src tests/"
run-demo = "python src/demo.py"
run-web = "streamlit run src/streamlit/web_interface.py"
ruff-check = "ruff check src tests"
ruff-fix = "ruff check --fix src tests"

[tool.poe.tasks.full-lint]
sequence = [
    { cmd = "black --check ." },
    { cmd = "isort --check ." },
    { cmd = "ruff check src tests" },
    { cmd = "mypy src/form_filler" }
]
help = "Run full linting suite without modifying files"

[tool.poe.tasks.fix-all]
sequence = [
    { cmd = "black ." },
    { cmd = "isort ." },
    { cmd = "ruff check --fix src tests" }
]
help = "Apply all automatic fixes (black, isort, ruff)"

[tool.poe.tasks.install]
sequence = [
    { cmd = "uv pip install -e \".\"" }
]
help = "Install the package in development mode"


[tool.poe.tasks.install-web]
sequence = [
    { cmd = "uv pip install -e \".[web]\"" }
]
help = "Install the package with web interface dependencies"

[tool.poe.tasks.install-dev]
sequence = [
    { cmd = "uv pip install -e \".[dev]\"" }
]
help = "Install the package with development dependencies"

[tool.poe.tasks.install-hooks]
sequence = [
    { cmd = "./scripts/install_hooks.sh" }
]
help = "Install pre-commit hooks in the current virtualenv"

[tool.poe.tasks.setup-secrets-detection]
sequence = [
    { cmd = "./scripts/setup_gitleaks_config.sh" }
]
help = "Set up custom gitleaks configuration for enhanced secrets detection"

[tool.poe.tasks.force-setup-secrets-detection]
sequence = [
    { cmd = "./scripts/setup_gitleaks_config.sh --force" }
]
help = "Set up custom gitleaks configuration, overwriting any existing config"

[tool.poe.tasks.install-all]
sequence = [
    { cmd = "uv pip install -e \".[dev,web,monitoring]\"" }
]
help = "Install the package with all dependencies"

[tool.poe.tasks.clean]
sequence = [
    { cmd = "rm -rf build/ dist/ *.egg-info/ .pytest_cache/ .coverage htmlcov/ .mypy_cache/" },
    { cmd = "find . -type d -name __pycache__ -exec rm -rf {} +; true" }
]
help = "Clean build artifacts and caches"

[tool.poe.tasks.version]
cmd = "python -c \"import form_filler; print(f'Form-filler version: {form_filler.__version__}')\""
help = "Show the current package version"

[tool.poe.tasks.serve]
cmd = "streamlit run src/streamlit/web_interface.py"
help = "Run the Streamlit web interface"
